% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acctsgeno.R
\name{gamfunction}
\alias{gamfunction}
\title{Accounting for Genotype Uncertainty using fractional counts}
\usage{
gamfunction(y0hat, y1hat, y2hat, y3hat, y4hat)
}
\arguments{
\item{Posterior}{probabilities}
}
\description{
Accounting for Genotype Uncertainty using fractional counts
}
