% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactposterior.R
\name{genotypelikelihood}
\alias{genotypelikelihood}
\title{Integrating over posterior genotypes}
\usage{
genotypelikelihood(gi0, gi1, gi2, gi3, gi4, B)
}
\arguments{
\item{gi0, ..}{Posterior probabilities}

\item{B}{Number of samples}
}
\value{
calculate the exact p value from this sample
}
\description{
Integrating over posterior genotypes
}
