% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfreq.R
\name{gfreq}
\alias{gfreq}
\title{Calculate genotype frequencies from gamete frequencies}
\usage{
gfreq(pvec)
}
\arguments{
\item{pvec}{A vector of length three (gamete frequencies) that sum to 1}
}
\value{
Genotype frequency.
}
\description{
These genotype frequencies are calculated  under the assumption of
random mating where each frequency is generated through a convolution
of the gamete frequencies.
}
\examples{
pvec <- c(0.1,0.4,0.5)
gfreq(pvec)
}
