% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisqr.R
\name{rmchisq}
\alias{rmchisq}
\title{Calculate Chi-squared for tetraploid}
\usage{
rmchisq(pvec, log_p = FALSE, N)
}
\arguments{
\item{pvec}{vector of length 3}

\item{log_p}{if the user desires to get the log of the p value}

\item{N}{the sample size}
}
\value{
the p-value
}
\description{
Calculate Chi-squared for tetraploid
}
