% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmexact-package.R
\docType{package}
\name{rmexact-package}
\alias{rmexact-package}
\alias{_PACKAGE}
\alias{rmexact}
\title{Exact tests for random mating in autopolyploids.}
\description{
The main functions are
\describe{
\item{\code{\link{tetexact}()}}{Exact test using sufficient statistics for random mating in autotetraploids when the genotypes are known.}
\item{\code{\link{rmslrt}()}}{Exact test using the split likelihood ratio approach of Wasserman et al. (2020) for random mating in autopolyploids when genotypes are known.}
\item{\code{\link{rmchisq}()}}{Chi-squared test for random mating in autopolyploids when genotypes are known.}
}
}
\references{
\itemize{
\item{Wasserman, L., Ramdas, A., & Balakrishnan, S. (2020). Universal inference. \emph{Proceedings of the National Academy of Sciences}, 117(29), 16880-16890. \doi{10.1073/pnas.1922664117}}
}
}
\author{
David Gerard and Karene Matoka Nana
}
\keyword{internal}
