% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmslrt.R
\name{rmslrt}
\alias{rmslrt}
\title{Split likelihood ratio test for random mating}
\usage{
rmslrt(nvec, sprop = 0.5, log_p = FALSE, nrep = 10)
}
\arguments{
\item{nvec}{The vector of genotype counts}

\item{sprop}{The proportional split.}

\item{log_p}{Should we return the log p-value?}

\item{nrep}{Number of iterations}
}
\value{
the p-value or the log of the p-value
}
\description{
Implements the split likelihood ratio test of Wasserman et al. (2020
when testing for random mating in autopolyploids.
}
\examples{
\dontrun{
set.seed(1)
p <- c(1/3, 1/3, 1/3)
q <- stats::convolve(p, rev(p), type = "open")
nvec <- c(stats::rmultinom(n = 1, size = 100, prob = q))
rmslrt(nvec, sprop = 0.5, nrep = 100)

q <- c(0.25, 0.2, 0.1, 0.2, 0.25)
nvec <- c(stats::rmultinom(n = 1, size = 100, prob = q))
rmslrt(nvec, sprop = 0.5, nrep = 100)
}

}
\references{
\itemize{
\item{Wasserman, L., Ramdas, A., & Balakrishnan, S. (2020). Universal inference. \emph{Proceedings of the National Academy of Sciences}, 117(29), 16880-16890. \doi{10.1073/pnas.1922664117}}
}
}
