% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact.R
\name{tetexact}
\alias{tetexact}
\title{Exact test for tetraploid random mating}
\usage{
tetexact(y, log_p = FALSE, frac = FALSE)
}
\arguments{
\item{y}{The vector of genotype counts (of of length 5). \code{y[k]} is
the number of individuals with genotype \code{k-1}.}

\item{log_p}{A logical. Should we return the log p-value (\code{TRUE}) or not
(\code{FALSE})?}

\item{frac}{A logical. Does \code{y} contain fractional genotypes (e.g.
because \code{y} is calculated by summing up posterior probabilities)?}
}
\description{
Conditions on sufficient statistics to calculate an exact p-value
against the null of random mating.
}
\details{
If you have genotype likelihoods, I would first convert those to genotype
posteriors (e.g. using \code{ldsep::gl_to_gp()}) to obtain \code{gp}, then
run \code{y <- colSums(gp, na.rm = TRUE)}. Then use the \code{frac = TRUE}?
}
\examples{
set.seed(1)

# Generate data according to random mating
p <- c(0.5, 0.3, 0.2)
q <- stats::convolve(p, rev(p), type = "open")
y <- c(stats::rmultinom(n = 1, size = 200, prob = q))
tetexact(y = y)

# Generate data different from random mating
q <- c(0.4, 0.3, 0.2, 0.1, 0.0)
y <- c(stats::rmultinom(n = 1, size = 200, prob = q))
tetexact(y = y)

## small change when using fractional counts
y <- y + runif(5)
tetexact(y = y)

}
\author{
David Gerard and Karene Matoka Nana
}
