% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact.R
\name{tetgp}
\alias{tetgp}
\title{Fuzzy p-values for random mating using genotype likelihoods.}
\usage{
tetgp(gp, nrep = 1000, log_p = FALSE)
}
\arguments{
\item{gp}{The matrix of genotype posteriors. \code{gp[i,k]} is the
posterior probability that individual \code{i} has genotype
\code{k - 1}.}

\item{nrep}{The number of type to run \code{\link{tetexact}()}.}

\item{log_p}{A logical. Should we return the log p-value (\code{TRUE}) or not
(\code{FALSE})?}
}
\value{
A sample from the distribution of the abstract randomized p-value.
}
\description{
Returns samples from the distribution of p-values. This distribution
is collectively called the "fuzzy p-value" (Geyer and Meeden, 2005).
}
\examples{
\dontrun{
set.seed(99)

# Generate data according to random mating
p <- c(0.5, 0.3, 0.2)
q <- stats::convolve(p, rev(p), type = "open")
y <- c(stats::rmultinom(n = 1, size = 1000, prob = q))
gp <- hwep::simgl(nvec = y, rdepth = 1000, ret = "gp")
pval1 <- tetgp(gp)
hist(pval1, breaks = 20, xlim = c(0, 1)) ## no evidence against null
vpv(pval1) ## vacuumed p-value

# Generate data different from random mating
q <- c(0.4, 0.3, 0.2, 0.1, 0.0)
y <- c(stats::rmultinom(n = 1, size = 100, prob = q))
gp <- hwep::simgl(nvec = y, rdepth = 1000, ret = "gp")
pval2 <- tetgp(gp)
hist(pval2, breaks = 20, xlim = c(0, 1)) ## evidence against null
vpv(pval2) ## vacuumed p-value
}

}
\references{
\itemize{
\item{C. J. Geyer and G. D. Meeden. Fuzzy and Randomized Confidence Intervals and P-Values. \emph{Statistical Science}, 20(4):358 – 366, 2005. \doi{10.1214/088342305000000340}.}
}
}
\author{
David Gerard
}
