% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal.R
\name{universal}
\alias{universal}
\title{Universal conference}
\usage{
universal(nvec, log_pvalue = FALSE)
}
\description{
Universal conference
}
\examples{
set.seed(1)
pvec <- c(1/3, 1/3, 1/3)
n <- 100
freq <- rmexact::gfreq(pvec = pvec)
nvec <- c(rmultinom(n = 1, size = n, prob = freq))
nvec <- c(1000, 100, 0, 1000, 0)
universal(nvec = nvec)
}
