% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact.R
\name{vpv}
\alias{vpv}
\title{Vacuum the fuzzy p-values}
\usage{
vpv(pvals)
}
\arguments{
\item{pvals}{A vector of sampled abstract randomized p-values
(Geyer and Meeden, 2005).}
}
\value{
The vacuumed p-value (VP-value)
}
\description{
Vacuum the fuzzy p-values
}
\examples{
set.seed(1)
pvals <- runif(1000)
vpv(pvals)

}
\references{
\itemize{
\item{C. J. Geyer and G. D. Meeden. Fuzzy and Randomized Confidence Intervals and P-Values. \emph{Statistical Science}, 20(4):358 – 366, 2005. \doi{10.1214/088342305000000340}.}
}
}
\author{
David Gerard
}
