% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisqr.R
\name{rmchisq}
\alias{rmchisq}
\title{Calculate Chi-squared for tetraploid}
\usage{
rmchisq(nvec, log_pvalue = FALSE)
}
\arguments{
\item{nvec}{vector of length 5}

\item{log_pvalue}{if the user desires to get the log of the p value}
}
\value{
the p-value or the log of the p-value depending on log_pvalue
}
\description{
Calculate Chi-squared for tetraploid
}
\examples{
nvec <- c(2, 1, 2, 3, 0)
rmchisq(nvec, log_pvalue = FALSE)
hwep::rmlike(nvec = nvec, thresh = 0)$p_rm
}
