% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmexact-package.R
\docType{package}
\name{rmexact-package}
\alias{rmexact-package}
\alias{_PACKAGE}
\alias{rmexact}
\title{Exact tests for random mating in autopolyploids.}
\description{
Right now, we only have an implementation for tetraploids. But we can
account for genotype uncertainty through genotype posteriors.
}
\details{
The main functions are
\describe{
\item{\code{\link{tetexact}()}}{Calculate the exact p-value when genotypes are known.}
\item{\code{\link{tetgp}()}}{Calculate the fuzzy p-value when using genotype posteriors.}
\item{\code{\link{vpv}()}}{Calculate the vacuumed p-value.}
}
}
\author{
David Gerard and Karene Matoka Nana
}
\keyword{internal}
